me.mwisbest.music !
This is based on the AOSP Music app. Yes, the thing that looks like it's for
Froyo (which it is). It has been heavily reworked to use newer APIs and
is being updated to look more like a proper Material app. Currently, it's more
like Materiyo than Material, but it's getting there. A lot of old cruft has
been removed from the code at least, and lots of bugs fixed.

Mission Statement: KISS (Keep It Simple, Stupid). The vast majority of music
apps have simply gone too far and over the top, eventually becoming laggy,
unusable, annoying messes.